post/migrations/0006_auto_20150302_1059.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import models, migrations
import sorl.thumbnail.fields


class Migration(migrations.Migration):

    dependencies = [
        ('post', '0005_post_photo'),
    ]

    operations = [
        migrations.AlterField(
            model_name='post',
            name='photo',
            field=sorl.thumbnail.fields.ImageField(upload_to=b'photos'),
            preserve_default=True,
        ),
    ]
