requirements.txt
Django==1.7.5
Pillow==2.7.0
argparse==1.2.1
django-twitter-bootstrap==3.3.0
djangorestframework==3.0.5
sorl-thumbnail==12.2
wsgiref==0.1.2
